+++ src/index.ts
@@ -2,6 +2,7 @@ export { doctorConfig, formatDoctor } from "./doctor";
 export { findConfigPath, loadConfig } from "./config";
 export { formatDecision } from "./format";
 export { starterConfig, writeStarterConfig } from "./init";
+export { formatPlan, planFreelane } from "./plan";
 export { formatProviderList, listProviders } from "./provider-list";
 export { getRunnerOption, providerFactories } from "./providers";
 export { displayUnit, quotaFor, roundQuota } from "./quota";